Fix the license identifier for GPL-3.0-or-later
Change the value of the license key to represent "GNU General Public
License v3.0 or later." Snapcraft supports expressions in SPDX
version 2.1, but this version defines only the deprecated identifier
'GPL-3.0+' for the license. SPDX version 2.2 adds the identifier
'GPL-3.0-or-later', but Snapcraft fails to recognize it. Change the
value of the license key to the deprecated identifier for now.

Wikipedia states:

  Starting version 2.0, it is no longer valid to use the '+' operator
  in a license identifier. By removing this syntax, it left an
  undefined state for licenses accepting the current version and those
  after it, such as the GPL. It was valid to use 'GPL-3.0-or-later',
  but it wasn't explicitly written in the specifications. This was
  fixed later with version 2.2.