Sync with hello-javafx and shorten Maven group IDs
Minimize the differences between this project and the hello-javafx
project in the Makefile, pom.xml, and snapcraft.yaml files.

Make the Maven groupId strings as short as possible. Give the parent
project and sub-projects more meaningful names.

Create the file 'snap/local/maven.yaml', an alternative build of the
Snap package that uses the core18 Maven plugin instead of the core20
Make plugin. The Make build uses jlink to create a minimal custom
runtime image, while the Maven build includes the entire Java Runtime
Environment. The sizes of each package and its contents are:

  Plugin  Package (MB)  Contents (MB)
  ------  ------------  -------------
  Make         25             86
  Maven        59            204