Add architectures, fix icon, and other changes
Change the Makefile default to use Ubuntu OpenJDK 11 from the Main
repository instead of the old OpenJDK 14 from the Universe repository.

Build the Snap package on all possible architectures for which the
'core20' and 'openjdk' Snap packages are both available.

Make the default 'hello-java' command run the Java console application
instead of the Swing application.

Get the source from the local project containing the 'snapcraft.yaml'
file instead of from the GitHub repository.

Delete empty directories only if they're also found in the base snap.

Add a white outline to the Duke icon image so that it looks good when
displayed on either a light or a dark background.