Put transparency back into the desktop icon image
Put the transparent background back into the desktop icon image of
Duke waving. Add a white outline to the foreground so that it looks
good when displayed on either a light or a dark background.

Revert the default 'hello-java' command back to the Java Swing
application so that it uses 'hello-java.png' for its desktop icon
instead of the system's default Java icon.